+++ packages/instrumentation-net/test/tls.test.ts
@@ -21,6 +21,7 @@ import {
 } from '@opentelemetry/sdk-trace-base';
 import { NodeTracerProvider } from '@opentelemetry/sdk-trace-node';
 import { AsyncLocalStorageContextManager } from '@opentelemetry/context-async-hooks';
+import { SemconvStability } from '@opentelemetry/instrumentation';
 import * as assert from 'assert';
 import * as tls from 'tls';
 import { NetInstrumentation } from '../src';
@@ -32,6 +33,10 @@ import {
   TLS_SERVER_KEY,
   PORT,
 } from './utils';
+
+// By default tests run with both old and stable semconv.
+process.env.OTEL_SEMCONV_STABILITY_OPT_IN = 'http/dup';
+const DEFAULT_NET_SEMCONV_STABILITY = SemconvStability.DUPLICATE;
 
 const memoryExporter = new InMemorySpanExporter();
 const provider = new NodeTracerProvider({
@@ -95,7 +100,11 @@ describe('NetInstrumentation', () => {
           },
         },
         () => {
-          assertTLSSpan(getTLSSpans(), tlsSocket);
+          assertTLSSpan(
+            getTLSSpans(),
+            tlsSocket,
+            DEFAULT_NET_SEMCONV_STABILITY
+          );
           done();
         }
       );
@@ -112,7 +121,7 @@ describe('NetInstrumentation', () => {
         c.end();
       });
       tlsSocket.once('end', () => {
-        assertTLSSpan(getTLSSpans(), tlsSocket);
+        assertTLSSpan(getTLSSpans(), tlsSocket, DEFAULT_NET_SEMCONV_STABILITY);
         done();
       });
     });
@@ -128,7 +137,11 @@ describe('NetInstrumentation', () => {
           },
         },
         () => {
-          assertTLSSpan(getTLSSpans(), tlsSocket);
+          assertTLSSpan(
+            getTLSSpans(),
+            tlsSocket,
+            DEFAULT_NET_SEMCONV_STABILITY
+          );
           done();
         }
       );