+++ packages/host-metrics/CHANGELOG.md
@@ -1,6 +1,27 @@
 <!-- markdownlint-disable MD007 MD034 -->
 # Changelog
 
+## [0.38.2](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/host-metrics-v0.38.1...host-metrics-v0.38.2) (2026-01-21)
+
+
+### Bug Fixes
+
+* **deps:** update dependency systeminformation to v5.30.3 ([#3335](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/3335)) ([9af0086](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/9af008622326464fa3076127bbab34a1dc0d10aa))
+
+## [0.38.1](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/host-metrics-v0.38.0...host-metrics-v0.38.1) (2026-01-14)
+
+
+### Bug Fixes
+
+* **deps:** update dependency systeminformation to v5.27.14 [security] ([#3308](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/3308)) ([c2d0bc5](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/c2d0bc5b1928e2147fe44c1a5a8b826f30b8f7de))
+
+## [0.38.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/host-metrics-v0.37.0...host-metrics-v0.38.0) (2025-12-17)
+
+
+### Features
+
+* apply rule @typescript-eslint/explicit-member-accessibility ([#3250](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/3250)) ([d0f9491](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/d0f94913cfd7fccc29053a0dc6f6aef6572235cc))
+
 ## [0.37.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/host-metrics-v0.36.2...host-metrics-v0.37.0) (2025-11-06)